heat: add cleanup_stacks
If you have a cloud where some of your stacks are stuck in
DELETE_FAILED state and you need to run something to clean
up all the resources, this tool is designed for this.

At the moment, it just handles nested stacks but it can be
extended to handle more things (i.e. network ports, etc). It
does come with destructive behaviour so the operator needs
to be careful using it.

Change-Id: I5c419aa0f45d243b585c9794406086891bfe06c1