Add br-int-flows-analyze.py script
This is a (somewhat poorly executed) attempt at making sense of the
flows dumped by ovs-ofctl dump-flows br-int.

Our idea was to have a tool that can help us filter flow rules
pertaining to a VM and, if possible, to make them more readable.

This tool can be used by running, for example :
br-int-flows-analyze.py --vm-uuid VM_UUID --ingress

It will first display the list of rules that MAY be traversed by ingress
traffic for the designated VM and after that, a summary with possibly
some annotations about the rules (for example, when a port range is
split across multiple rules, it will try to merge them together).

Change-Id: I251e5b4c719d32c6283eb0e8888e7510dd35cea2